telegram bot/requirements.txt
fastapi 
uvicorn
requests
python-dotenv
huggingface_hub
